return true + unique chainreverseresolver
