+++ deploy/dnssec-oracle/00_deploy_algorithms.ts
@@ -8,6 +8,8 @@ const func: DeployFunction = async function (hre) {
   await viem.deploy('P256SHA256Algorithm', [])
 
   if (network.tags.test) await viem.deploy('DummyAlgorithm', [])
+
+  return true
 }
 
 func.id = 'dnssec-algorithms v1.0.0'
