+++ deploy/dnssec-oracle/10_deploy_oracle.ts
@@ -118,6 +118,8 @@ const func: DeployFunction = async function (hre) {
   await Promise.all(
     transactions.map(async (hash) => viem.waitForTransactionSuccess(hash)),
   )
+
+  return true
 }
 
 func.id = 'DNSSECImpl v1.0.0'
