+++ deploy/l2/00_deploy_l2_reverse_registrar.ts
@@ -312,6 +312,8 @@ const func: DeployFunction = async function (hre) {
 
     await viem.deploy('L2ReverseRegistrar', [coinType])
   }
+
+  return true
 }
 
 func.id = 'L2ReverseRegistrar v1.0.0'
